code/requirements.txt
pillow == 11.3.0
tqdm == 4.67.1
diffusers == 0.34.0
transformers == 4.53.2
pandas == 2.3.1
matplotlib == 3.10.3
scikit-learn == 1.7.0
faiss-cpu == 1.11.0
accelerate == 1.9.0
torch_fidelity == 0.3.0
nudenet == 3.4.2
prettytable == 3.16.0

